+++ src/star-rating.vue
@@ -1,8 +1,8 @@
 <template>
     <div :class="['vue-star-rating', {'vue-star-rating-rtl':rtl}, {'vue-star-rating-inline': inline}]">
         <div @mouseleave="resetRating" class="vue-star-rating">
-            <span v-for="n in maxRating" :key="n" :class="[{'vue-star-rating-pointer': !readOnly }, 'vue-star-rating-star']">
-              <star :fill="fillLevel[n-1]" :size="starSize" :points="starPoints" :star-id="n" :step="step" :active-color="activeColor" :inactive-color="inactiveColor" :border-color="borderColor" :border-width="borderWidth" :rounded-corners="roundedCorners" :padding="padding" @star-selected="setRating($event, true)" @star-mouse-move="setRating" :rtl="rtl" :glow="glow" :glow-color="glowColor"></star>
+            <span v-for="n in maxRating" :key="n" :class="[{'vue-star-rating-pointer': !readOnly }, 'vue-star-rating-star']" :style="{'margin-right': margin + 'px'}">
+              <star :fill="fillLevel[n-1]" :size="starSize" :points="starPoints" :star-id="n" :step="step" :active-color="activeColor" :inactive-color="inactiveColor" :border-color="borderColor" :border-width="borderWidth" :rounded-corners="roundedCorners" @star-selected="setRating($event, true)" @star-mouse-move="setRating" :rtl="rtl" :glow="glow" :glow-color="glowColor"></star>
             </span>
             <span v-if="showRating" :class="['vue-star-rating-rating-text', textClass]"> {{formattedRating}}</span>
         </div>
@@ -94,11 +94,11 @@ export default {
             type: Number,
             default: null
         },
-        glow:{
+        glow: {
             type: Number,
             default: 0
         },
-        glowColor:{
+        glowColor: {
             type: String,
             default: '#fff'
         }
@@ -108,7 +108,6 @@ export default {
         this.currentRating = this.rating
         this.selectedRating = this.currentRating
         this.createStars(this.roundStartRating)
-
     },
     methods: {
         setRating($event, persist) {
@@ -155,6 +154,9 @@ export default {
         },
         shouldRound() {
             return this.ratingSelected || this.roundStartRating
+        },
+        margin() {
+            return this.padding + this.borderWidth
         }
     },
     watch: {